Replace Label synonym with String in some parsers
The Label type synonym (which is used to identify labels in assembly),
is used as a plane String in cInstrDest, cInstrJump and cInstrAluOps
parsers. This doesn't cause any compilation or run-time problems because
Label is just a type synonym for String, but is ugly and breaks the
intended meaning of the Label type.

Signed-off-by: Collin J. Doering <collin.doering@rekahsoft.ca>